+++ video_xblock/__init__.py
@@ -2,7 +2,7 @@
 Video xblock module.
 """
 
-__version__ = '1.1.0'
+__version__ = '1.1.1'
 
 # pylint: disable=wildcard-import
 from .video_xblock import *  # nopep8